Give all thread pool executors a unique name for Espresso idle resources
This fixes the final issue with instability in test results while using
background threads with Espresso testing. Hilt creates a new
application and component tree for every test, but old thread pool
executors are not getting shutdown leaving them registered as an
Idling Resource with Espresso. Since they all use the same name, only
the first one is ever added and later tests in the same run will not
be registered causing them to proceed on before background threads are
complete. This solution is not ideal, but it is now 100% reliable
across a range of API levels.