Added in theme support from the upstream FragmentScenario
With this change, the last remaining test is now fixed. The default
theme that was being selected did not include the Content
Description in the TimePicker sub-fields with some tests rely on.
With this change, the HiltFragmentScenario behaves exacting like
the upstream version which it was copied from except that it
supports Hilt.

All tests are now passing on this branch and it is now more
reliable than main due to the use of Espresso idling resources.